% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_grid.R
\name{build_grid}
\alias{build_grid}
\title{Title}
\usage{
build_grid(
  node_locs,
  center_lat,
  center_lon,
  x_size_meters,
  y_size_meters,
  bin_size
)
}
\arguments{
\item{node_locs}{dataframe of node locations}

\item{center_lat}{grid center latitude}

\item{center_lon}{grid centerlongitude}

\item{x_size_meters}{grid size x-coord, in meters}

\item{y_size_meters}{grid size y-coord, in meters}

\item{bin_size}{number of bins in grid}
}
\description{
Title
}
