% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_grid_values.R
\name{calc_grid_values}
\alias{calc_grid_values}
\title{Calculate Grid Values}
\usage{
calc_grid_values(grid_df, rec_df, rssi_coefs)
}
\arguments{
\item{grid_df}{grid dataframe}

\item{rec_df}{recorded detections dataframe}

\item{rssi_coefs}{RSSI vs Distance fit coefficients from calibration}
}
\value{
grid_value_data_frame
}
\description{
Calculate Grid Values
}
\examples{
calc_grid_values(grid_df, rec_df, rssi_coefs)
}
