% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_rssi_v_dist.R
\name{calc_rssi_v_dist}
\alias{calc_rssi_v_dist}
\title{Title}
\usage{
calc_rssi_v_dist(node_locs, sidekick_tag_df, detection_df, use_sync = TRUE)
}
\arguments{
\item{use_sync}{}
}
\description{
Title
}
