% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_tag_activity.R
\name{calculate_tag_activity}
\alias{calculate_tag_activity}
\title{Title}
\usage{
calculate_tag_activity(tag_dets, tag_beep_interval)
}
\arguments{
\item{tag_beep_interval}{}
}
\description{
Title
}
