% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_track.R
\name{calculate_track}
\alias{calculate_track}
\title{Calculate Track}
\usage{
calculate_track(
  start_time,
  length_seconds,
  step_size_seconds,
  det_time_window,
  filter_alpha,
  filter_time_range,
  grid_df,
  detection_df,
  node_locs,
  node_t_offset = NULL,
  rssi_coefs,
  track_frame_output_path = NULL,
  tile_url = "https://tile.openstreetmap.org/{z}/{x}/{y}.png"
)
}
\arguments{
\item{start_time}{POSIXct datetime value, in UTC/GMT timezone}

\item{length_seconds}{Time (s)}

\item{step_size_seconds}{Step Size (s)}

\item{det_time_window}{Detection time window (s)}

\item{filter_alpha}{Alpha value, used in calc_receiver_values function}

\item{filter_time_range}{Time range (s) to include detections in filtered value}

\item{grid_df}{Grid dataframe}

\item{detection_df}{Detections dataframe}

\item{node_locs}{Node locations}

\item{node_t_offset}{Node time offset (obtained in Ch. 5)}

\item{rssi_coefs}{RSSI vs Distance fit coefficients from calibration}

\item{track_frame_output_path}{Filepath to save track map}

\item{tile_url}{URL for map}
}
\value{
track_df
}
\description{
Calculate Track
}
\examples{
track_df <- calculate_track(
                            start_time = "2023-08-03 19:50:45",
                            length_seconds = 1050,
                            step_size_seconds = 10,
                            det_time_window = 60, # Must have detection within this window to be included in position calculation
                            filter_alpha = 0.7,
                            filter_time_range = 120, # Time range to include detections in filtered value
                            grid_df = grid_df,
                            detection_df = detection_df,
                            node_locs = node_locs,
                            node_t_offset = node_toff_df,
                            rssi_coefs = rssi_coefs,
                            track_frame_output_path = NULL # If NULL no individual frames will be saved
                            )
}
