% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celltracktech-package.R
\docType{package}
\name{celltracktech-package}
\alias{celltracktech}
\alias{celltracktech-package}
\title{celltracktech: Cellular Tracking Technologies}
\description{
Tools for analyzing data from CTT products.
}
\author{
\strong{Maintainer}: Jessica Gorzo \email{jessica.gorzo@celltracktech.com} (\href{https://orcid.org/0000-0002-3715-8229}{ORCID})

}
\keyword{internal}
