% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classic_plot_theme.R
\docType{data}
\name{classic_plot_theme}
\alias{classic_plot_theme}
\title{Title}
\format{
An object of class \code{theme} (inherits from \code{gg}) of length 10.
}
\usage{
classic_plot_theme()
}
\description{
Title
}
\keyword{datasets}
