% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_db.R
\name{create_db}
\alias{create_db}
\title{Create Postgres database}
\usage{
create_db(conn)
}
\arguments{
\item{conn}{the connection to your local database}
}
\description{
This function allows you to create a blank version of the CTT designed database
}
\examples{
create_db(conn)
}
