% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_postgres.R
\name{create_node_db}
\alias{create_node_db}
\title{Create database from Node Data files from SD card}
\usage{
create_node_db(my_token, outpath, myproject, db_name)
}
\arguments{
\item{outpath}{where your files are to be downloaded}

\item{myproject}{the name of your project on our system}

\item{db_name}{the connection to your local database}
}
\description{
Create database from Node Data files from SD card
}
\examples{
create_node_db(outpath = './data/', myproject = 'Meadows V2', db_name = conn)
}
