% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_postgres.R
\name{db_insert}
\alias{db_insert}
\title{Title}
\usage{
db_insert(contents, filetype, conn, sensor = NA, y, begin = NULL)
}
\arguments{
\item{contents}{Dataframe}

\item{filetype}{filetype}

\item{conn}{database connection}

\item{sensor}{NA}

\item{y}{file path}

\item{begin}{date station was deployed}
}
\description{
Title
}
