% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{deployments}
\alias{deployments}
\title{Meadows Deployment Data}
\format{
\subsection{\code{deployments} A data frame with 125 rows and 5 columns:}{

\describe{
\item{TagId}{Tag ID}
\item{DeployDate}{Date tag was deployed in the field}
\item{Species}{Alpha-numeric code for species}
\item{TagType}{Tag Type: Power, Hybride, Life, Blu}
\item{AntennaType}{1/8 wave or 1/4 wave}
}
}
}
\usage{
deployments
}
\description{
Tag deployment data for the Meadows V2 project
}
\keyword{datasets}
