% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detection_summary.R
\name{detection_summary}
\alias{detection_summary}
\title{Detection Summary}
\usage{
detection_summary(detection_df, tag_list)
}
\arguments{
\item{detection_df}{Tag detection dataframe}

\item{tag_list}{Tag IDs from deployment list}
}
\value{
det_summary
}
\description{
Detection Summary
}
\examples{
detection_summary(detection_df, tag_list)

}
