% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_grid.R
\name{draw_grid}
\alias{draw_grid}
\title{Map Grid}
\usage{
draw_grid(node_locs, grid_df)
}
\arguments{
\item{node_locs}{Node locations dataframe}

\item{grid_df}{Grid dataframe}
}
\value{
map
}
\description{
Map Grid
}
\examples{
grid_map <- draw_grid(node_locs, grid_df)
}
