% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_single_solution.R
\name{draw_single_solution}
\alias{draw_single_solution}
\title{Draw Single Solution}
\usage{
draw_single_solution(rec_df, grid_values, solution, multilat_sol, tile_url)
}
\arguments{
\item{rec_df}{Dataframe produced by calc_receiver_values function}

\item{grid_values}{Grid values dataframe produced by calc_grid_values function}

\item{solution}{Subset of test_grid_values for max test_grid_values$value}

\item{multilat_sol}{Subset of multilat_fit coefficients}

\item{tile_url}{Map URL}
}
\value{
map - map with grid
}
\description{
Draw Single Solution
}
\examples{
test_map <- draw_single_solution(test_rec_df,
                                 test_grid_values,
                                 solution,
                                 multilat_result,
                                 my_tile_url)




}
