% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_files.R
\name{error_files}
\alias{error_files}
\title{Find files with errors}
\usage{
error_files(dirin, dirout, conn = NULL)
}
\arguments{
\item{dirin}{where your files were downloaded}

\item{dirout}{where you want the files with errors to be copied}
}
\description{
This function allows you to copy files with errors or corrupt lines into a new destination to look closer
}
\examples{
error_files("~/mydata", "~/errorfiles")
}
