% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.distance.R
\name{estimate.distance}
\alias{estimate.distance}
\alias{estimate_distance}
\title{Estimate Distance}
\usage{
estimate.distance(x, K, a, S)

estimate_distance(...)
}
\arguments{
\item{x}{combined_data dataframe}

\item{K}{horizontal asymptote}

\item{a}{intercept}

\item{S}{decay factor}
}
\value{
combined.data
}
\description{
Estimate Distance
}
\examples{
combined_data <- estimate_distance(combined_data, K, a, S)
}
