% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filecatch.R
\name{findfiles}
\alias{findfiles}
\title{Catch files with errors}
\usage{
findfiles(outpath, dest)
}
\arguments{
\item{outpath}{where your files were downloaded}

\item{dest}{where flawed files will go}
}
\description{
This function allows you to find files that have erroneous lines
}
\examples{
findfiles("~/mydata", "~/error_files")
}
