% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_my_data.R
\name{get_my_data}
\alias{get_my_data}
\title{Download data}
\usage{
get_my_data(
  my_token,
  outpath,
  db_name = NULL,
  myproject = NULL,
  mystation = NULL,
  begin = NULL,
  end = NULL,
  filetypes = NULL
)
}
\arguments{
\item{my_token}{your API key}

\item{outpath}{where your files are to be downloaded}

\item{db_name}{(optional) the connection to your local database}

\item{myproject}{the name of your project on our system}

\item{mystation}{(optional) the station ID you'd like to download data from}

\item{begin}{(optional) limit your data download to a start time}

\item{end}{(optional) limit your data download to an end time}
}
\description{
This function allows you to download your sensor station data, with the option of simultaneously importing to a local database
}
\examples{
get_my_data(token, "~/mydata", myproject = "Project Name from CTT Account")
}
