% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_my_projects.R
\name{get_my_projects}
\alias{get_my_projects}
\title{Get list of projects associated with your account}
\usage{
get_my_projects(my_token)
}
\arguments{
\item{my_token}{API authentication token}
}
\value{
character vector of project names
}
\description{
Get list of projects associated with your account
}
