% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tag_detections.R
\name{get_tag_detection_count}
\alias{get_tag_detection_count}
\title{Get Tag Detections}
\usage{
get_tag_detection_count(detection_df, min_det_count)
}
\arguments{
\item{detection_df}{Tag detection dataframe}

\item{min_det_count}{Minimum number of detections to include}
}
\value{
tag_counts - dataframe
}
\description{
Get Tag Detections
}
\examples{
get_tag_detection_count(detection_df, min_det_count)


}
