% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_time_value.R
\name{get_time_value}
\alias{get_time_value}
\title{Get Time Value}
\usage{
get_time_value(time_string)
}
\arguments{
\item{time_string}{time_utc column in Sidekick dataframe}
}
\value{
time_value
}
\description{
Get Time Value
}
\examples{
get_time_value(sidekick_df$time_utc)

}
