% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newdb.R
\name{import_node_data}
\alias{import_node_data}
\title{Incorporate node data}
\usage{
import_node_data(d, outpath, myproject = NULL, station_id)
}
\arguments{
\item{d}{The connection to your local database}

\item{outpath}{The overarching directory where your "nodes" folder lives, within your project folder (if applicable)}

\item{myproject}{(optional) If you have a sensor station related to this project, place your "nodes" folder in the folder named for your project as it
appears on the CTT website. If you use the API to pull sensor station data, this folder has been automatically created}
}
\description{
This function allows you to include data pulled directly from your nodes. Create a folder called "nodes" and create sub-folders named for each node ID.
Put the beep files pulled from the node(s) in their appropriate folder(s)
}
\examples{
conn <- DBI::dbConnect(RPostgres::Postgres(), dbname="mydb")
outpath <- "~/Documents/my_project/"
list.files(outpath)
# "My Project"
import_node_data(conn, outpath, myproject="My Project")
}
