% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_csv.R
\name{load_csv}
\alias{load_csv}
\title{Function to load CSV or CSV.GZ file}
\usage{
load_csv(file_path)
}
\arguments{
\item{file_path}{\itemize{
\item string
}}
}
\value{
data
}
\description{
Function to load CSV or CSV.GZ file
}
\examples{
load_csv('./data/')
load_csv('./data/blu-file.csv')
load_csv('./data/blu-file.csv.gz')
}
