% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newdb.R
\name{load_node_data}
\alias{load_node_data}
\title{Load Node Data}
\usage{
load_node_data(e, conn, outpath, myproject, station_id)
}
\arguments{
\item{e}{file path}

\item{conn}{database connection}

\item{outpath}{outpath directory}

\item{myproject}{project name}

\item{station_id}{sensor station id - string}
}
\value{
NULL on success, or a list containing error details
}
\description{
Load Node Data
}
\examples{
\dontrun{
load_node_data(e, conn, outpath, myproject, station_id)
}
}
