% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_node_health.R
\name{load_node_health_files}
\alias{load_node_health_files}
\title{Load Node Health}
\usage{
load_node_health_files(directory, start_time = NULL, stop_time = NULL)
}
\arguments{
\item{directory}{Directory that contains node health files}

\item{start_time}{POSIXct datetime, in UTC}

\item{stop_time}{POSIXct datetime, in UTC}
}
\value{
df
}
\description{
Load Node Health
}
\examples{
\dontrun{
load_node_health_files("path/to/health_data", start_time, stop_time)
}
}
