% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_sidekick_data.R
\name{load_sidekick_data}
\alias{load_sidekick_data}
\title{Load Sidekick Data}
\usage{
load_sidekick_data(sidekick_file_path)
}
\arguments{
\item{sidekick_file_path}{Directory that contains sidekick data files}
}
\value{
df
}
\description{
Load Sidekick Data
}
\examples{
sidekick_all_df <- load_sidekick_data(sidekick_file_path)
}
