% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_calibration_track.R
\name{map_calibration_track}
\alias{map_calibration_track}
\title{Map Calibration Track}
\usage{
map_calibration_track(
  node_locs,
  sidekick_tag_df,
  tile_url = "https://tile.openstreetmap.org/{z}/{x}/{y}.png"
)
}
\arguments{
\item{node_locs}{Node locations dataframe}

\item{sidekick_tag_df}{Sidekick tag dataframe of a specific tag ID}

\item{tile_url}{Map URL}
}
\value{
map
}
\description{
Map Calibration Track
}
\examples{
calibration_map <- map_calibration_track(node_locs,
                                         sidekick_tag_df,
                                         tile_url = my_tile_url)
}
