% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_grid_solution.R
\name{map_latest_solution}
\alias{map_latest_solution}
\title{Map Latest Solution}
\usage{
map_latest_solution(
  node_locs,
  rec_df,
  grid_values,
  solution,
  track_df,
  tile_url
)
}
\arguments{
\item{node_locs}{Node location dataframe}

\item{rec_df}{Receiver dataframe}

\item{grid_values}{Grid values dataframe}

\item{solution}{Subset of test_grid dataframe}

\item{track_df}{Tracks dataframe}

\item{tile_url}{Map URL}
}
\value{
map
}
\description{
Map Latest Solution
}
\examples{
\dontrun{
map_latest_solution(node_locs, rec_df, grid_values, solution, track_df, tile_url)
}
}
