% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_w_trilat.R
\name{map_multilat}
\alias{map_multilat}
\title{Compare with Trilateration Results}
\usage{
map_multilat(
  node_locs,
  track_error_df,
  sidekick_df,
  multilat_df,
  tile_url = "https://tile.openstreetmap.org/{z}/{x}/{y}.png"
)
}
\arguments{
\item{node_locs}{Node locations dataframe}

\item{track_error_df}{Track error dataframe}

\item{sidekick_df}{Sidekick detections dataframe}

\item{multilat_df}{Multilateration dataframe}

\item{tile_url}{Map URL}
}
\value{
map
}
\description{
Compare with Trilateration Results
}
\examples{
\dontrun{
map_multilat(node_locs, track_error_df, sidekick_df, multilat_df)
}
}
