% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_node_locations.R
\name{map_node_locations}
\alias{map_node_locations}
\title{Map Node Locations}
\usage{
map_node_locations(
  node_locs,
  tile_url = "https://tile.openstreetmap.org/{z}/{x}/{y}.png"
)
}
\arguments{
\item{node_locs}{\itemize{
\item data frame
}}

\item{tile_url}{\itemize{
\item url to map
}}
}
\value{
\itemize{
\item map
}
}
\description{
Map Node Locations
}
\examples{
node_map <- map_node_locations(node_locs, tile_url = my_tile_url)
}
