% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_track.R
\name{map_track}
\alias{map_track}
\title{Map Track}
\usage{
map_track(
  node_locs,
  track_df,
  tile_url = "https://tile.openstreetmap.org/{z}/{x}/{y}.png"
)
}
\arguments{
\item{node_locs}{Node locations database}

\item{track_df}{Dataframe produced by calculate_track function}

\item{tile_url}{Map URL}
}
\value{
map
}
\description{
Map Track
}
\examples{
track_map <- map_track(node_locs, track_df, my_tile_url)
}
