% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node.R, R/node_file.R
\name{node_file}
\alias{node_file}
\title{Make the node file}
\usage{
node_file(health)

node_file(health)
}
\arguments{
\item{health}{}
}
\description{
This function allows you to create the node location spatial data frame from your node health data.
}
