% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/out.R
\name{out}
\alias{out}
\title{Out}
\usage{
out(x, contents, timezone)
}
\arguments{
\item{x}{Dataframe}

\item{contents}{Dataframe containing the data}

\item{timezone}{Timezone, UTC}
}
\value{
newtimecol
}
\description{
Out
}
\examples{
\dontrun{
out("time", contents, "UTC")
}
}
