% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseit.R
\name{parseit}
\alias{parseit}
\title{Function to parse Payload and extract temperature and battery voltage}
\usage{
parseit(y)
}
\arguments{
\item{y}{\itemize{
\item hexstring
}}
}
\value{
vector - solar battery voltage, temperature in C
}
\description{
Function to parse Payload and extract temperature and battery voltage
}
\examples{
parsed <- t(sapply(blu[[payload_column]], parseit))
}
