% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patch.R
\name{patch}
\alias{patch}
\title{Fix database entries}
\usage{
patch(d, outpath, myproject, dirout)
}
\arguments{
\item{d}{database connection}

\item{outpath}{where your files were downloaded}

\item{myproject}{the name of your project at CTT}

\item{dirout}{where the problem files will be copied}
}
\description{
This function allows you to copy files with errors or corrupt lines into a new destination to look closer
}
\examples{
patch(dbConnect(RPostgres::Postgres(), dbname = db_name), "~/mydata", "My Project", "~/errfiles")
}
