% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_calibration_result.R
\name{plot_calibration_result}
\alias{plot_calibration_result}
\title{Plot Calibration Result}
\usage{
plot_calibration_result(rssi_v_dist, theme = NULL)
}
\arguments{
\item{rssi_v_dist}{Dataframe produced by calc_rssi_v_dist function}

\item{theme}{ggplot2 theme}
}
\value{
plot
}
\description{
Plot Calibration Result
}
\examples{
calibration_plot <- plot_calibration_result(rssi_v_dist, classic_plot_theme())
}
