% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_node_locations.R
\name{plot_node_locations}
\alias{plot_node_locations}
\title{Plot Node Locations}
\usage{
plot_node_locations(node_health_df, node_locs, theme = NULL)
}
\arguments{
\item{theme}{}
}
\value{
plot
}
\description{
Plot Node Locations
}
