% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rssi_v_dist.R
\name{predict_dist}
\alias{predict_dist}
\title{Predict Distance}
\usage{
predict_dist(rssi_coefs, rssi)
}
\arguments{
\item{rssi_coefs}{Numeric vector of RSSI model coefficients (a, b, c)}

\item{rssi}{RSSI value}
}
\value{
rssi_coefs
}
\description{
Predict Distance
}
\examples{
\dontrun{
predict_dist(c(-95, 60, 0.05), -80)
}
}
