% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rssi_v_dist.R
\name{predict_rssi}
\alias{predict_rssi}
\title{RSSI Coefficients}
\usage{
predict_rssi(rssi_coefs, dist)
}
\arguments{
\item{rssi_coefs}{Numeric vector of RSSI model coefficients (a, b, c)}

\item{dist}{Distance value}
}
\value{
rssi_coefs
}
\description{
RSSI Coefficients
}
\examples{
\dontrun{
predict_rssi(c(-95, 60, 0.05), 100)
}
}
