% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_directory.R
\name{process_directory}
\alias{process_directory}
\title{Main function to process all CSV/CSV.GZ files in a directory}
\usage{
process_directory(input_dir, output_dir)
}
\arguments{
\item{output_dir}{}
}
\description{
Main function to process all CSV/CSV.GZ files in a directory
}
\examples{
process_directory('./data/', './data/results/')
}
