% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_file.R
\name{process_file}
\alias{process_file}
\title{Function to process a single file}
\usage{
process_file(file_path, output_dir)
}
\arguments{
\item{file_path}{filename}

\item{output_dir}{output directory}
}
\value{
csv with parsed battery voltage and temperature
}
\description{
Function to process a single file
}
\examples{
process_file('./blu-data.csv', './')
}
