% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sidekick_all_df}
\alias{sidekick_all_df}
\title{Another sidekick calibration file}
\format{
\subsection{\code{sidekick_all_df}}{

A data frame with 7,240 rows and 60 columns:
}

A data frame with 716 rows and 8 columns:
\describe{
\item{tag_type}{Tag type detected}
\item{tag_id}{Tag ID}
\item{time_utc}{Time in UTC}
\item{rssi}{Received signal strength indicator in dBm}
\item{lat}{Latitude}
\item{lon}{Longitude}
\item{heading}{Cardinal direction}
\item{antenna_angle}{Angle of antenna}
}
}
\usage{
sidekick_all_df
}
\description{
A subset of data from the World Health Organization Global Tuberculosis
Report ...
}
\keyword{datasets}
