% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sidekick_cal}
\alias{sidekick_cal}
\title{Sidekick Calibration Data}
\format{
A data frame with 716 rows and 8 columns:
\describe{
\item{tag_type}{Tag type detected}
\item{tag_id}{Tag ID}
\item{time_utc}{Time in UTC}
\item{rssi}{Received signal strength indicator in dBm}
\item{lat}{Latitude}
\item{lon}{Longitude}
\item{heading}{Cardinal direction}
\item{antenna_angle}{Angle of antenna}
}
}
\usage{
sidekick_cal
}
\description{
An example of the sidekick calibration in the Meadows V2 field site.
}
\keyword{datasets}
