% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sidekick_cal_test2}
\alias{sidekick_cal_test2}
\title{Sidekick Calibration Test Data 2}
\format{
\subsection{\code{sidekick_cal_test2} A data frame with 415 rows and 8 columns:}{

\describe{
\item{tag_type}{Tag type detected}
\item{tag_id}{Tag ID}
\item{time_utc}{Time in UTC}
\item{rssi}{Received signal strength indicator in dBm}
\item{lat}{Latitude}
\item{lon}{Longitude}
\item{heading}{Cardinal direction}
\item{antenna_angle}{Angle of antenna}
}
}
}
\usage{
sidekick_cal_test2
}
\description{
An example of the sidekick test calibration in the Meadows V2 field site.
}
\keyword{datasets}
