% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tag_hist_plot_theme.R
\name{tag_hist_plot_theme}
\alias{tag_hist_plot_theme}
\title{Tag Histogram Plot Theme}
\usage{
tag_hist_plot_theme()
}
\value{
plot
}
\description{
Tag Histogram Plot Theme
}
\examples{
ggplot(node_record_counts,
       aes(x = factor(x = node_id,
               levels = node_id),
           y = n)) +
  geom_bar(stat = "identity") +
  coord_flip() +
  labs(x = "Health Record Count",
       y = "Node Id") +
  tag_hist_plot_theme()
}
