% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trilateration.R
\name{trilateration}
\alias{trilateration}
\title{Trilateration}
\usage{
trilateration(x, nodes, RSS.FILTER)
}
\arguments{
\item{x}{Dataframe with tag detections (raw or blu), produced by prep_data function}

\item{nodes}{Nodes dataframe}

\item{RSS.FILTER}{Filter for RSS}
}
\value{
estimated.location_results
}
\description{
Trilateration
}
\examples{
\dontrun{
trilateration(x, nodes, RSS.FILTER)
}
}
