% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trilateration.TestData.Distance.Filter.R
\name{trilateration.TestData.Distance.Filter}
\alias{trilateration.TestData.Distance.Filter}
\alias{trilateration_testdata_distance_filter}
\title{Trilateration - Test Data Distance Filter}
\usage{
trilateration.TestData.Distance.Filter(x, DIST.FILTER)

trilateration_testdata_distance_filter(...)
}
\arguments{
\item{x}{\itemize{
\item dataframe
}}

\item{DIST.FILTER}{List ex. c(315,500,750,1000)}
}
\value{
list
}
\description{
Trilateration - Test Data Distance Filter
}
\examples{
Dist.filters <- trilateration_testdata_distance_filter(combined_data, c(315,500,750,1000))
}
