% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_existing_parallel.R
\name{update_existing_parallel}
\alias{update_existing_parallel}
\title{Update existing blu and node_blu database tables - using parallel}
\usage{
update_existing_parallel(table_name, con)
}
\arguments{
\item{table_name}{string}

\item{payload}{dataframe column}

\item{id}{dataframe column}
}
\description{
Update existing blu and node_blu database tables - using parallel
}
\examples{
\dontrun{
update_existing_parallel("blu", con)
}
}
