% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_avg_activity.R
\name{calc_avg_activity}
\alias{calc_avg_activity}
\title{Title}
\usage{
calc_avg_activity(tag_activity, start_time, stop_time)
}
\arguments{
\item{stop_time}{}
}
\value{
result
}
\description{
Title
}
\examples{
calc_avg_activity(tag_activity, start_time, stop_time)

}
