% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_location_density.R
\name{calc_location_density}
\alias{calc_location_density}
\title{Calculate Location Density}
\usage{
calc_location_density(grid_df, locations_df)
}
\arguments{
\item{grid_df}{Grid dataframe; created by build_grid function}

\item{locations_df}{Locations dataframe; created by calculate_track function}
}
\value{
loc_density_df
}
\description{
Calculate Location Density
}
\examples{
calc_location_density(grid_df, locations_df)
}
