% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_track_error.R
\name{calc_track_error}
\alias{calc_track_error}
\title{Calculate Track Errors}
\usage{
calc_track_error(sidekick_df, track_df)
}
\arguments{
\item{sidekick_df}{Sidekick tag detection dataframe}

\item{track_df}{Tracks dataframe, obtained from calculate_track function}
}
\value{
solution_compare_df
}
\description{
Calculate Track Errors
}
\examples{
calc_track_error(sidekick_df, track_df)
}
