% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_track.R
\name{calculate_track}
\alias{calculate_track}
\title{Title}
\usage{
calculate_track(
  start_time,
  length_seconds,
  step_size_seconds,
  det_time_window,
  filter_alpha,
  filter_time_range,
  grid_df,
  detection_df,
  node_locs,
  node_t_offset = NULL,
  rssi_coefs,
  track_frame_output_path = NULL,
  tile_url = "https://tile.openstreetmap.org/{z}/{x}/{y}.png"
)
}
\arguments{
\item{tile_url}{}
}
\description{
Title
}
