% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classic_plot_theme.R
\name{classic_plot_theme}
\alias{classic_plot_theme}
\title{Classic Plot Theme}
\usage{
classic_plot_theme()
}
\value{
creates classic theme for ggplots
}
\description{
Classic Plot Theme
}
\examples{
node_loc_plot <- plot_node_locations(node_health_df,
                                     node_locs,
                                     theme = classic_plot_theme())
}
