% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_outpath.R
\name{create_outpath}
\alias{create_outpath}
\title{Title}
\usage{
create_outpath(outpath)
}
\arguments{
\item{outpath}{a string of the directory or nested directories where you want to save your data}
}
\value{
creates directory/directories in your current working directory
}
\description{
Title
}
\examples{
create_outpath('./data/')
create_outpath('./data/meadows/')
}
