% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.setup.R
\name{data.setup}
\alias{data.setup}
\title{Combine Detection Data and Sidekick Calibration Data}
\usage{
data.setup(
  test,
  testdata_in,
  nodes,
  tag_col,
  tagid,
  time_col,
  timezone,
  x,
  y,
  node_ids = NULL,
  loc_precision = NA,
  latlon = T,
  fileloc,
  filetype
)
}
\arguments{
\item{filetype}{}
}
\value{
combined.data - dataframe
}
\description{
Combine Detection Data and Sidekick Calibration Data
}
