% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_single_solution.R
\name{draw_single_solution}
\alias{draw_single_solution}
\title{Draw Single Solution}
\usage{
draw_single_solution(rec_df, grid_values, solution, multilat_sol, tile_url)
}
\arguments{
\item{tile_url}{}
}
\value{
map - map with grid
}
\description{
Draw Single Solution
}
