% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_node_locations.R
\name{export_node_locations}
\alias{export_node_locations}
\title{Export Node Locations}
\usage{
export_node_locations(output_file_path, node_locs)
}
\arguments{
\item{output_file_path}{.csv file name}

\item{node_locs}{dataframe}
}
\description{
Export Node Locations
}
