% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_node_detection_data.R
\name{load_node_detection_data}
\alias{load_node_detection_data}
\title{Load Node Detection Data}
\usage{
load_node_detection_data(directory, start_time = NULL, stop_time = NULL)
}
\arguments{
\item{directory}{Directory that contains node data}

\item{start_time}{POSIXct datetime in UTC}

\item{stop_time}{POSIXct datetime in UTC}
}
\value{
df
}
\description{
Load Node Detection Data
}
