% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_calibration_track.R
\name{map_calibration_track}
\alias{map_calibration_track}
\title{Map Calibration Track}
\usage{
map_calibration_track(
  node_locs,
  sidekick_tag_df,
  tile_url = "https://tile.openstreetmap.org/{z}/{x}/{y}.png"
)
}
\arguments{
\item{tile_url}{}
}
\value{
map
}
\description{
Map Calibration Track
}
