% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_location_density.R
\name{map_location_density}
\alias{map_location_density}
\title{Map Location Density}
\usage{
map_location_density(
  loc_density_df,
  tile_url = "https://tile.openstreetmap.org/{z}/{x}/{y}.png"
)
}
\arguments{
\item{loc_density_df}{Dataframe produced by calc_location_density function}

\item{tile_url}{Map URL}
}
\value{
map
}
\description{
Map Location Density
}
\examples{
loc_density_map <- map_location_density(loc_density_df = loc_density,
                                        my_tile_url)
}
