% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_location_density.R
\name{map_location_density}
\alias{map_location_density}
\title{Title}
\usage{
map_location_density(
  loc_density_df,
  tile_url = "https://tile.openstreetmap.org/{z}/{x}/{y}.png"
)
}
\arguments{
\item{tile_url}{}
}
\description{
Title
}
