% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_grid_solution.R
\name{map_single_solution}
\alias{map_single_solution}
\title{Map Single Solution}
\usage{
map_single_solution(node_locs, rec_df, grid_values, solution, tile_url)
}
\arguments{
\item{node_locs}{Node locations dataframe}

\item{rec_df}{Receiver dataframe}

\item{grid_values}{Grid values}

\item{solution}{Subset of test_grid_values for max test_grid_values$value}

\item{tile_url}{Map URL}
}
\value{
map
}
\description{
Map Single Solution
}
