% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_track.R
\name{map_track}
\alias{map_track}
\title{Title}
\usage{
map_track(
  node_locs,
  track_df,
  tile_url = "https://tile.openstreetmap.org/{z}/{x}/{y}.png"
)
}
\arguments{
\item{tile_url}{}
}
\description{
Title
}
