% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_track_error.R
\name{map_track_error}
\alias{map_track_error}
\title{Title}
\usage{
map_track_error(
  node_locs,
  track_error_df,
  sidekick_df,
  tile_url = "https://tile.openstreetmap.org/{z}/{x}/{y}.png"
)
}
\arguments{
\item{tile_url}{}
}
\description{
Title
}
