% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping.R
\name{mapping}
\alias{mapping}
\title{Mapping}
\usage{
mapping(
  node_locs,
  track_error_df,
  tile_url = "https://tile.openstreetmap.org/{z}/{x}/{y}.png"
)
}
\arguments{
\item{node_locs}{Node locations dataframe}

\item{track_error_df}{Dataframe produced by calc_track_error function}

\item{tile_url}{Map URL}
}
\value{
map
}
\description{
Mapping
}
\examples{
mapping(nodes, location.estimates)
}
