% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node.R
\name{node_file}
\alias{node_file}
\title{Make the node file}
\usage{
node_file(health)
}
\arguments{
\item{health}{a data frame of node health data pulled from your local database}
}
\description{
This function allows you to create the node location spatial data frame from your node health data.
}
