% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_node_locations.R
\name{plot_node_locations}
\alias{plot_node_locations}
\title{Plot Node Locations}
\usage{
plot_node_locations(node_health_df, node_locs, theme = NULL)
}
\arguments{
\item{node_health_df}{Node health dataframe}

\item{node_locs}{Node locations}

\item{theme}{ggplot2 theme}
}
\value{
plot
}
\description{
Plot Node Locations
}
\examples{
node_loc_plot <- plot_node_locations(node_health_df,
                                     node_locs,
                                     theme = classic_plot_theme())
}
