% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_postgres.R
\name{update_db}
\alias{update_db}
\title{Insert data into your local database}
\usage{
update_db(d, outpath, myproject, fix = FALSE)
}
\arguments{
\item{d}{the connection to your local database}

\item{outpath}{where your files were downloaded}

\item{myproject}{the name of your project on our system}

\item{fix}{(optional) if TRUE, it will re-check your files to see if any imports were missed}
}
\description{
This function allows you to import data you've downloaded to your database
}
\examples{
update_db(conn, "~/mydata", myproject="Project Name from CTT Account", fix=FALSE)
}
