% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_existing_blu.R
\name{update_existing_blu}
\alias{update_existing_blu}
\title{Update existing blu and node_blu database tables}
\usage{
update_existing_blu(table_name, con)
}
\arguments{
\item{table_name}{string}

\item{payload}{dataframe column}

\item{id}{dataframe column}
}
\description{
Update existing blu and node_blu database tables
}
